# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehouse Management API",
        "description": "API for warehouse management system including inventory, stock movements, attendance, and reporting.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{host}",
            "description": "API Server",
            "variables": {
                "host": {
                    "default": "http://localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/api/barang-masuk/export/excel": {
            "get": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "Export barang masuk to Excel",
                "operationId": "f80cca0d4187f89843386643d96ec481",
                "responses": {
                    "200": {
                        "description": "Excel file download"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-masuk/{barang_masuk}/print-surat-jalan": {
            "get": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "Print surat jalan for barang masuk",
                "operationId": "dadfdb86e8c2d7b2a0fdb323683ea6d9",
                "parameters": [
                    {
                        "name": "barang_masuk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF surat jalan"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-masuk": {
            "get": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "List all barang masuk",
                "operationId": "b7323e16e67422ef3067507c935ce828",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of barang masuk"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "Create barang masuk",
                "operationId": "c636b48db7f55fb4889a60d858645734",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBarangMasukRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barang masuk created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-masuk/{barang_masuk}": {
            "get": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "Get barang masuk by ID",
                "operationId": "aee6389b6e73bd5969060c5f6d29ece9",
                "parameters": [
                    {
                        "name": "barang_masuk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang masuk detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "Update barang masuk",
                "operationId": "ed2cb8a64203a5d3183e67687e4dd11a",
                "parameters": [
                    {
                        "name": "barang_masuk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBarangMasukRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang masuk updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barang Masuk"
                ],
                "summary": "Delete barang masuk",
                "operationId": "c068fbf662f316c14cbf6365e1b67e54",
                "parameters": [
                    {
                        "name": "barang_masuk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Barang masuk deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/mutasi-stok": {
            "get": {
                "tags": [
                    "Mutasi Stok"
                ],
                "summary": "List all mutasi stok",
                "operationId": "76d49141de429c6d8f2cfcd45a93e89a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of mutasi stok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mutasi Stok"
                ],
                "summary": "Create mutasi stok",
                "operationId": "7416c6c95b2649f1109fb6c6fe18141e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMutasiStokRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mutasi stok created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mutasi-stok/{mutasi_stok}": {
            "get": {
                "tags": [
                    "Mutasi Stok"
                ],
                "summary": "Get mutasi stok by ID",
                "operationId": "6f067a1f8d42f1c9656b3320aa51921e",
                "parameters": [
                    {
                        "name": "mutasi_stok",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mutasi stok detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mutasi Stok"
                ],
                "summary": "Update mutasi stok",
                "operationId": "2a840d3bbc901c90677e05b58fd9ceca",
                "parameters": [
                    {
                        "name": "mutasi_stok",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMutasiStokRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mutasi stok updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mutasi Stok"
                ],
                "summary": "Delete mutasi stok",
                "operationId": "dc8bf071505a37455dd1df50513a3668",
                "parameters": [
                    {
                        "name": "mutasi_stok",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mutasi stok deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreBarangMasukRequest": {
                "required": [
                    "no_referensi",
                    "gudang_id",
                    "supplier_id",
                    "tanggal"
                ],
                "properties": {
                    "no_referensi": {
                        "type": "string",
                        "example": "BM001"
                    },
                    "nomor_surat_jalan": {
                        "type": "string",
                        "nullable": true
                    },
                    "gudang_id": {
                        "type": "integer"
                    },
                    "supplier_id": {
                        "type": "integer"
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date"
                    },
                    "keterangan": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "default": "pending",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected",
                            "partial"
                        ]
                    },
                    "dokumen": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreMutasiStokRequest": {
                "required": [
                    "no_referensi",
                    "barang_id",
                    "gudang_asal_id",
                    "gudang_tujuan_id",
                    "qty",
                    "tanggal"
                ],
                "properties": {
                    "no_referensi": {
                        "type": "string",
                        "example": "MS001"
                    },
                    "barang_id": {
                        "type": "integer"
                    },
                    "gudang_asal_id": {
                        "type": "integer"
                    },
                    "gudang_tujuan_id": {
                        "type": "integer"
                    },
                    "lokasi_rak_asal_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "lokasi_rak_tujuan_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "qty": {
                        "type": "number",
                        "format": "float"
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date"
                    },
                    "keterangan": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "default": "pending",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected",
                            "completed"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Passport",
                "description": "Enter Bearer token obtained from login endpoint"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}